requirements.txt
tqdm
numpy
spacy
mysql-connector-python
flask
flask_cors
